refactor(cast): inline internal helpers implementation
