+++ crates/cast/src/args.rs
@@ -14,8 +14,8 @@ use alloy_eips::{Encodable2718, eip7702::SignedAuthorization};
 use alloy_ens::{NameOrAddress, ProviderEnsExt, namehash};
 use alloy_network::{BlockResponse, Ethereum, Network, eip2718::Decodable2718};
 use alloy_primitives::{
-    Address, B256, Bytes, I256, LogData, TxHash, U64, U256, b256, eip191_hash_message, hex,
-    keccak256,
+    Address, B256, Bytes, I256, Keccak256, LogData, TxHash, U64, U256, b256, eip191_hash_message,
+    hex, keccak256,
     utils::{ParseUnits, Unit},
 };
 use alloy_provider::Provider;
@@ -163,7 +163,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
             } else {
                 data
             };
-            let out = data.iter().map(|s| crate::strip_0x(s)).collect::<String>();
+            let out = data.iter().map(|s| strip_0x(s)).collect::<String>();
             print_scalar(format!("0x{out}"))?;
         }
         CastSubcommand::FromBin => {
@@ -250,7 +250,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
         }
         CastSubcommand::ToBytes32 { bytes } => {
             let s = stdin::unwrap_line(bytes)?;
-            let s = crate::strip_0x(&s);
+            let s = strip_0x(&s);
             if s.len() > 64 {
                 eyre::bail!("string >32 bytes");
             }
@@ -272,7 +272,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
         }
         CastSubcommand::Pad { data, right, left: _, len } => {
             let s = stdin::unwrap_line(data)?;
-            let s = crate::strip_0x(&s);
+            let s = strip_0x(&s);
             let hex_len = len
                 .checked_mul(2)
                 .filter(|&h| h <= u16::MAX as usize)
@@ -313,7 +313,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
         }
         CastSubcommand::ParseBytes32Address { bytes } => {
             let s = stdin::unwrap_line(bytes)?;
-            let s = crate::strip_0x(&s);
+            let s = strip_0x(&s);
             if s.len() != 64 {
                 eyre::bail!("expected 64 byte hex-string, got {s}");
             }
@@ -376,7 +376,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
             let mut data_tokens = Vec::new();
             for (input, token) in event.inputs.iter().zip(tokens) {
                 if input.indexed {
-                    topics.push(crate::encode_event_topic(&token));
+                    topics.push(encode_event_topic(&token));
                 } else {
                     data_tokens.push(token);
                 }
@@ -430,7 +430,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
                 let event = get_event(&event_sig)?;
                 event.decode_log_parts(core::iter::once(event.selector()), &hex::decode(data)?)?
             } else {
-                let data = crate::strip_0x(&data);
+                let data = strip_0x(&data);
                 let selector: B256 = data.get(..64).unwrap_or_default().parse()?;
                 let Some(event) = SignaturesIdentifier::new(false)?.identify_event(selector).await
                 else {
@@ -447,7 +447,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
             let error = if let Some(err_sig) = sig {
                 get_error(&err_sig)?
             } else {
-                let data = crate::strip_0x(&data);
+                let data = strip_0x(&data);
                 let selector = data.get(..8).unwrap_or_default();
                 let Some(error) =
                     SignaturesIdentifier::new(false)?.identify_error(selector.parse()?).await
@@ -978,7 +978,7 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
         CastSubcommand::MakeTx(cmd) => cmd.run().await?,
         CastSubcommand::PublishTx { raw_tx, cast_async, rpc } => {
             let provider = rpc_provider(&rpc)?;
-            let raw_tx = hex::decode(crate::strip_0x(&raw_tx))?;
+            let raw_tx = hex::decode(strip_0x(&raw_tx))?;
             let pending_tx = provider.send_raw_transaction(&raw_tx).await?;
             if cast_async {
                 print_scalar(format!("{:#x}", pending_tx.inner().tx_hash()))?;
@@ -1195,24 +1195,14 @@ pub async fn run_command(args: CastArgs) -> Result<()> {
             let config = etherscan.load_config()?;
             let chain = config.chain.unwrap_or_default();
             let api_key = config.get_etherscan_api_key(Some(chain));
+            let client = explorer_client(chain, api_key, explorer_api_url, explorer_url)?;
+            let metadata = client.contract_source_code(address.parse()?).await?;
             match (directory, flatten) {
                 (Some(dir), false) => {
-                    explorer_client(chain, api_key, explorer_api_url, explorer_url)?
-                        .contract_source_code(address.parse()?)
-                        .await?
-                        .source_tree()
-                        .write_to(&dir)?;
+                    metadata.source_tree().write_to(&dir)?;
                 }
-                (None, false) => sh_println!(
-                    "{}",
-                    explorer_client(chain, api_key, explorer_api_url, explorer_url)?
-                        .contract_source_code(address.parse()?)
-                        .await?
-                        .source_code()
-                )?,
+                (None, false) => sh_println!("{}", metadata.source_code())?,
                 (dir, true) => {
-                    let client = explorer_client(chain, api_key, explorer_api_url, explorer_url)?;
-                    let metadata = client.contract_source_code(address.parse()?).await?;
                     let Some(metadata) = metadata.items.first() else {
                         eyre::bail!("Empty contract source code");
                     };
@@ -1532,4 +1522,132 @@ fn get_selector(signature: &str, optimize: usize) -> Result<(String, String)> {
             None
         })
         .ok_or_eyre("No selector found")
+}
+
+fn strip_0x(s: &str) -> &str {
+    s.strip_prefix("0x").unwrap_or(s)
+}
+
+/// Encodes the topic of an indexed event parameter.
+///
+/// Value types are encoded as their 32-byte word. Reference types are hashed over the special
+/// in-place encoding defined for indexed event parameters, which differs from regular ABI
+/// encoding: `string` and `bytes` contribute their raw contents, and array or struct members are
+/// concatenated recursively without any offsets or length prefixes.
+///
+/// See <https://docs.soliditylang.org/en/latest/abi-spec.html#encoding-of-indexed-event-parameters>
+pub(super) fn encode_event_topic(value: &DynSolValue) -> B256 {
+    if let Some(word) = value.as_word() {
+        return word;
+    }
+    // Top-level `string` and `bytes` hash their raw contents without padding.
+    if let Some(bytes) = value.as_packed_seq() {
+        return keccak256(bytes);
+    }
+    let mut preimage = Vec::new();
+    encode_event_topic_preimage(value, &mut preimage);
+    keccak256(preimage)
+}
+
+/// Encodes a value into the in-place preimage of an indexed event parameter: words as-is,
+/// `string`/`bytes` right-padded to a multiple of 32 bytes, and sequences as the concatenation of
+/// their encoded members.
+fn encode_event_topic_preimage(value: &DynSolValue, out: &mut Vec<u8>) {
+    if let Some(word) = value.as_word() {
+        out.extend_from_slice(word.as_slice());
+    } else if let Some(bytes) = value.as_packed_seq() {
+        let pad = bytes.len().next_multiple_of(32) - bytes.len();
+        out.extend_from_slice(bytes);
+        out.resize(out.len() + pad, 0);
+    } else if let Some(values) = value.as_fixed_seq().or_else(|| value.as_array()) {
+        for value in values {
+            encode_event_topic_preimage(value, out);
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use alloy_primitives::{Address, Bytes, U256};
+    use alloy_sol_types::{EventTopic, sol_data};
+
+    /// Compares [`super::encode_event_topic`] against alloy's static [`EventTopic`]
+    /// implementation, which `sol!`-generated events use to compute indexed topics.
+    #[test]
+    fn encode_event_topic_matches_static_encoding() {
+        let uint = |n: u64| DynSolValue::Uint(U256::from(n), 256);
+        let string = |s: &str| DynSolValue::String(s.into());
+        let topic = |v: &DynSolValue| super::encode_event_topic(v);
+
+        let long = "abcdefghijklmnopqrstuvwxyz0123456789abcd";
+        for s in ["", "hello", long] {
+            assert_eq!(
+                topic(&string(s)),
+                <sol_data::String as EventTopic>::encode_topic(&s.to_string()).0,
+                "string {s:?}"
+            );
+        }
+
+        let bytes = hex::decode("deadbeef").unwrap();
+        assert_eq!(
+            topic(&DynSolValue::Bytes(bytes.clone())),
+            <sol_data::Bytes as EventTopic>::encode_topic(&Bytes::from(bytes)).0,
+        );
+
+        let addr = Address::repeat_byte(0x42);
+        assert_eq!(
+            topic(&DynSolValue::Address(addr)),
+            <sol_data::Address as EventTopic>::encode_topic(&addr).0,
+        );
+
+        assert_eq!(
+            topic(&DynSolValue::Array(vec![uint(1), uint(2)])),
+            <sol_data::Array<sol_data::Uint<256>> as EventTopic>::encode_topic(&vec![
+                U256::from(1),
+                U256::from(2)
+            ])
+            .0,
+        );
+
+        assert_eq!(
+            topic(&DynSolValue::FixedArray(vec![uint(7), uint(9)])),
+            <sol_data::FixedArray<sol_data::Uint<256>, 2> as EventTopic>::encode_topic(&[
+                U256::from(7),
+                U256::from(9)
+            ])
+            .0,
+        );
+
+        assert_eq!(
+            topic(&DynSolValue::Array(vec![string("alpha"), string(long)])),
+            <sol_data::Array<sol_data::String> as EventTopic>::encode_topic(&vec![
+                "alpha".to_string(),
+                long.to_string()
+            ])
+            .0,
+        );
+
+        assert_eq!(
+            topic(&DynSolValue::Tuple(vec![uint(7), string("hello")])),
+            <(sol_data::Uint<256>, sol_data::String) as EventTopic>::encode_topic(&(
+                U256::from(7),
+                "hello".to_string()
+            ))
+            .0,
+        );
+
+        assert_eq!(
+            topic(&DynSolValue::Array(vec![
+                DynSolValue::Array(vec![uint(1)]),
+                DynSolValue::Array(vec![uint(2), uint(3)]),
+            ])),
+            <sol_data::Array<sol_data::Array<sol_data::Uint<256>>> as EventTopic>::encode_topic(
+                &vec![vec![U256::from(1)], vec![U256::from(2), U256::from(3)]]
+            )
+            .0,
+        );
+    }
+
+    // <https://github.com/foundry-rs/foundry/issues/2681>
 }
